% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastictranscoder_operations.R
\name{elastictranscoder_read_pipeline}
\alias{elastictranscoder_read_pipeline}
\title{The ReadPipeline operation gets detailed information about a pipeline}
\usage{
elastictranscoder_read_pipeline(Id)
}
\arguments{
\item{Id}{[required] The identifier of the pipeline to read.}
}
\description{
The ReadPipeline operation gets detailed information about a pipeline.
}
\section{Request syntax}{
\preformatted{elastictranscoder$read_pipeline(
  Id = "string"
)
}
}

\keyword{internal}
