% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesisvideomedia_service.R
\name{kinesisvideomedia}
\alias{kinesisvideomedia}
\title{Amazon Kinesis Video Streams Media}
\usage{
kinesisvideomedia()
}
\description{
Amazon Kinesis Video Streams Media
}
\section{Operations}{

\tabular{ll}{
\link[=kinesisvideomedia_get_media]{get_media} \tab Use this API to retrieve media content from a Kinesis video stream
}
}

\examples{
\donttest{svc$get_media(
  Foo = 123
)}

}
