% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_untag_resource}
\alias{mediaconnect_untag_resource}
\title{Deletes the specified tags from a resource}
\usage{
mediaconnect_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to remove tags from.}

\item{TagKeys}{[required] The keys of the tags to be removed.}
}
\description{
Deletes the specified tags from a resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
