% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconvert_operations.R
\name{mediaconvert_associate_certificate}
\alias{mediaconvert_associate_certificate}
\title{Associates an AWS Certificate Manager (ACM) Amazon Resource Name (ARN)
with AWS Elemental MediaConvert}
\usage{
mediaconvert_associate_certificate(Arn)
}
\arguments{
\item{Arn}{[required] The ARN of the ACM certificate that you want to associate with your MediaConvert resource.}
}
\description{
Associates an AWS Certificate Manager (ACM) Amazon Resource Name (ARN) with AWS Elemental MediaConvert.
}
\section{Request syntax}{
\preformatted{svc$associate_certificate(
  Arn = "string"
)
}
}

\keyword{internal}
