% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconvert_operations.R
\name{mediaconvert_delete_job_template}
\alias{mediaconvert_delete_job_template}
\title{Permanently delete a job template you have created}
\usage{
mediaconvert_delete_job_template(Name)
}
\arguments{
\item{Name}{[required] The name of the job template to be deleted.}
}
\description{
Permanently delete a job template you have created.
}
\section{Request syntax}{
\preformatted{mediaconvert$delete_job_template(
  Name = "string"
)
}
}

\keyword{internal}
