% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconvert_operations.R
\name{mediaconvert_describe_endpoints}
\alias{mediaconvert_describe_endpoints}
\title{Send an request with an empty body to the regional API endpoint to get
your account API endpoint}
\usage{
mediaconvert_describe_endpoints(MaxResults, Mode, NextToken)
}
\arguments{
\item{MaxResults}{Optional. Max number of endpoints, up to twenty, that will be returned at one time.}

\item{Mode}{Optional field, defaults to DEFAULT. Specify DEFAULT for this operation to return your endpoints if any exist, or to create an endpoint for you and return it if one doesn't already exist. Specify GET_ONLY to return your endpoints if any exist, or an empty list if none exist.}

\item{NextToken}{Use this string, provided with the response to a previous request, to request the next batch of endpoints.}
}
\description{
Send an request with an empty body to the regional API endpoint to get your account API endpoint.
}
\section{Request syntax}{
\preformatted{svc$describe_endpoints(
  MaxResults = 123,
  Mode = "DEFAULT"|"GET_ONLY",
  NextToken = "string"
)
}
}

\keyword{internal}
