% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconvert_operations.R
\name{mediaconvert_get_preset}
\alias{mediaconvert_get_preset}
\title{Retrieve the JSON for a specific preset}
\usage{
mediaconvert_get_preset(Name)
}
\arguments{
\item{Name}{[required] The name of the preset.}
}
\description{
Retrieve the JSON for a specific preset.
}
\section{Request syntax}{
\preformatted{mediaconvert$get_preset(
  Name = "string"
)
}
}

\keyword{internal}
