% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconvert_operations.R
\name{mediaconvert_get_queue}
\alias{mediaconvert_get_queue}
\title{Retrieve the JSON for a specific queue}
\usage{
mediaconvert_get_queue(Name)
}
\arguments{
\item{Name}{[required] The name of the queue that you want information about.}
}
\description{
Retrieve the JSON for a specific queue.
}
\section{Request syntax}{
\preformatted{svc$get_queue(
  Name = "string"
)
}
}

\keyword{internal}
