% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconvert_operations.R
\name{mediaconvert_list_presets}
\alias{mediaconvert_list_presets}
\title{Retrieve a JSON array of up to twenty of your presets}
\usage{
mediaconvert_list_presets(Category, ListBy, MaxResults, NextToken,
  Order)
}
\arguments{
\item{Category}{Optionally, specify a preset category to limit responses to only presets from that category.}

\item{ListBy}{Optional. When you request a list of presets, you can choose to list them alphabetically by NAME or chronologically by CREATION_DATE. If you don't specify, the service will list them by name.}

\item{MaxResults}{Optional. Number of presets, up to twenty, that will be returned at one time}

\item{NextToken}{Use this string, provided with the response to a previous request, to request the next batch of presets.}

\item{Order}{When you request lists of resources, you can optionally specify whether they are sorted in ASCENDING or DESCENDING order. Default varies by resource.}
}
\description{
Retrieve a JSON array of up to twenty of your presets. This will return the presets themselves, not just a list of them. To retrieve the next twenty presets, use the nextToken string returned with the array.
}
\section{Request syntax}{
\preformatted{mediaconvert$list_presets(
  Category = "string",
  ListBy = "NAME"|"CREATION_DATE"|"SYSTEM",
  MaxResults = 123,
  NextToken = "string",
  Order = "ASCENDING"|"DESCENDING"
)
}
}

\keyword{internal}
