% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_describe_input_security_group}
\alias{medialive_describe_input_security_group}
\title{Produces a summary of an Input Security Group}
\usage{
medialive_describe_input_security_group(InputSecurityGroupId)
}
\arguments{
\item{InputSecurityGroupId}{[required] The id of the Input Security Group to describe}
}
\description{
Produces a summary of an Input Security Group
}
\section{Request syntax}{
\preformatted{svc$describe_input_security_group(
  InputSecurityGroupId = "string"
)
}
}

\keyword{internal}
