% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_describe_offering}
\alias{medialive_describe_offering}
\title{Get details for an offering}
\usage{
medialive_describe_offering(OfferingId)
}
\arguments{
\item{OfferingId}{[required] Unique offering ID, e.g. '87654321'}
}
\description{
Get details for an offering.
}
\section{Request syntax}{
\preformatted{svc$describe_offering(
  OfferingId = "string"
)
}
}

\keyword{internal}
