% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_list_input_security_groups}
\alias{medialive_list_input_security_groups}
\title{Produces a list of Input Security Groups for an account}
\usage{
medialive_list_input_security_groups(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Produces a list of Input Security Groups for an account
}
\section{Request syntax}{
\preformatted{medialive$list_input_security_groups(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
