% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_start_channel}
\alias{medialive_start_channel}
\title{Starts an existing channel}
\usage{
medialive_start_channel(ChannelId)
}
\arguments{
\item{ChannelId}{[required] A request to start a channel}
}
\description{
Starts an existing channel
}
\section{Request syntax}{
\preformatted{svc$start_channel(
  ChannelId = "string"
)
}
}

\keyword{internal}
