% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_list_tags_for_resource}
\alias{mediapackage_list_tags_for_resource}
\title{List tags for resource}
\usage{
mediapackage_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required]}
}
\description{
List tags for resource
}
\section{Request syntax}{
\preformatted{mediapackage$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
