% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_rotate_channel_credentials}
\alias{mediapackage_rotate_channel_credentials}
\title{Changes the Channel's first IngestEndpoint's username and password}
\usage{
mediapackage_rotate_channel_credentials(Id)
}
\arguments{
\item{Id}{[required] The ID of the channel to update.}
}
\description{
Changes the Channel's first IngestEndpoint's username and password. WARNING - This API is deprecated. Please use RotateIngestEndpointCredentials instead
}
\section{Request syntax}{
\preformatted{svc$rotate_channel_credentials(
  Id = "string"
)
}
}

\keyword{internal}
