% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_rotate_ingest_endpoint_credentials}
\alias{mediapackage_rotate_ingest_endpoint_credentials}
\title{Rotate the IngestEndpoint's username and password, as specified by the
IngestEndpoint's id}
\usage{
mediapackage_rotate_ingest_endpoint_credentials(Id, IngestEndpointId)
}
\arguments{
\item{Id}{[required] The ID of the channel the IngestEndpoint is on.}

\item{IngestEndpointId}{[required] The id of the IngestEndpoint whose credentials should be rotated}
}
\description{
Rotate the IngestEndpoint's username and password, as specified by the IngestEndpoint's id.
}
\section{Request syntax}{
\preformatted{svc$rotate_ingest_endpoint_credentials(
  Id = "string",
  IngestEndpointId = "string"
)
}
}

\keyword{internal}
