% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapackage_operations.R
\name{mediapackage_update_channel}
\alias{mediapackage_update_channel}
\title{Updates an existing Channel}
\usage{
mediapackage_update_channel(Description, Id)
}
\arguments{
\item{Description}{A short text description of the Channel.}

\item{Id}{[required] The ID of the Channel to update.}
}
\description{
Updates an existing Channel.
}
\section{Request syntax}{
\preformatted{svc$update_channel(
  Description = "string",
  Id = "string"
)
}
}

\keyword{internal}
