% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatailor_service.R
\name{mediatailor}
\alias{mediatailor}
\title{AWS MediaTailor}
\usage{
mediatailor()
}
\description{
Use the AWS Elemental MediaTailor SDK to configure scalable ad insertion
for your live and VOD content. With AWS Elemental MediaTailor, you can
serve targeted ads to viewers while maintaining broadcast quality in
over-the-top (OTT) video applications. For information about using the
service, including detailed information about the settings covered in
this guide, see the AWS Elemental MediaTailor User Guide.

Through the SDK, you manage AWS Elemental MediaTailor configurations the
same as you do through the console. For example, you specify ad
insertion behavior and mapping information for the origin server and the
ad decision server (ADS).
}
\section{Operations}{

\tabular{ll}{
\link[=mediatailor_delete_playback_configuration]{delete_playback_configuration} \tab Deletes the playback configuration for the specified name \cr
\link[=mediatailor_get_playback_configuration]{get_playback_configuration} \tab Returns the playback configuration for the specified name \cr
\link[=mediatailor_list_playback_configurations]{list_playback_configurations} \tab Returns a list of the playback configurations defined in AWS Elemental MediaTailor \cr
\link[=mediatailor_list_tags_for_resource]{list_tags_for_resource} \tab Returns a list of the tags assigned to the specified playback configuration resource\cr
\link[=mediatailor_put_playback_configuration]{put_playback_configuration} \tab Adds a new playback configuration to AWS Elemental MediaTailor \cr
\link[=mediatailor_tag_resource]{tag_resource} \tab Adds tags to the specified playback configuration resource \cr
\link[=mediatailor_untag_resource]{untag_resource} \tab Removes tags from the specified playback configuration resource
}
}

\examples{
\donttest{svc$delete_playback_configuration(
  Foo = 123
)}

}
